update DHHH.edmx, MOD.sql, MoodleUser,AccountController, JSONController, _Layout.cshtml add jquery.blockUI.js

+++ trunk/CongThongTinSV/SP_SinhVien_Result.cs
@@ -20,6 +20,6 @@ namespace CongThongTinSV
         public Nullable<System.DateTime> Ngay_sinh { get; set; }
         public string Gioi_tinh { get; set; }
         public string Ten_lop { get; set; }
-        public Nullable<long> ID_moodle { get; set; }
+        public Nullable<int> ID_moodle { get; set; }
     }
 }
